push: follow source code pattern as other cmds
Move run(..) into push's create, as is done for all other commands.
